replace checkOwnerhsip function by searching contacts by userId property directrly in DB

+++ services/contactsService.js
@@ -4,12 +4,12 @@ async function getContacts(currUser) {
   return await Contact.find({ owner: currUser });
 }
 
-async function getContactById(contactId) {
-  return await Contact.findOne({ _id: contactId });
+async function getContactById(userId, contactId) {
+  return await Contact.findOne({ _id: contactId, owner: userId });
 }
 
-async function removeContact(contactId) {
-  return await Contact.findOneAndDelete({ _id: contactId });
+async function removeContact(userId, contactId) {
+  return await Contact.findOneAndDelete({ _id: contactId, owner: userId });
 }
 
 async function addContact(reqBody) {
@@ -19,10 +19,14 @@ async function addContact(reqBody) {
   return result;
 }
 
-async function updateContact(contactId, body) {
-  const updatedContact = Contact.findByIdAndUpdate({ _id: contactId }, body, {
-    returnDocument: "after",
-  });
+async function updateContact(userId, contactId, body) {
+  const updatedContact = Contact.findByIdAndUpdate(
+    { _id: contactId, owner: userId },
+    body,
+    {
+      returnDocument: "after",
+    }
+  );
   return updatedContact;
 }
